Cross-Encoder Re-Ranking
A second approach to re-ranking is to use a cross-encoder. Our semantic
search embeddings were created with a bi-encoder, which embeds queries
and documents separately so you can use cosine similarity to get a
relevance score.A cross-encoder does the same thing but embeds the query
and document as a single input to produce a relevance score directly.
The output is the score, no need to compute cosine
similarity.Cross-encoders see the full context of how the query and
document interact, so they can catch subtle relationships that
bi-encoders miss.Re-Ranking SpeedOne big advantage of cross-encoders is
that they are much faster and cheaper than LLMs, as they do not require
the same level of computational resources. Instead of using a tool with
all the capabilities of an LLM, this type of re-ranker can only do 1
thing: Take in a query and document pair and output a relevance score.In
other words, it's a regression model.Re-Ranking SpeedOne big advantage
of cross-encoders is that they are much faster and cheaper than LLMs, as
they do not require the same level of computational resources. Instead
of using a tool with all the capabilities of an LLM, this type of
re-ranker can only do 1 thing: Take in a query and document pair and
output a relevance score.In other words, it's a regression model.